requirements.txt
# Core dependencies
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Document processing
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
mammoth==1.6.0

# Evaluation metrics
bert-score==0.3.13
rouge-score==0.1.2
nltk==3.8.1
scikit-learn==1.3.2

# Data processing
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
requests==2.31.0
tqdm==4.66.1
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
